Removes animate that almost never did anything
This animate wouldn't run if the jump was too big, meaning we were only ever animating 1 or 2 pixels of parallax, and i do not believe the cost on the server nor the client is warranted for how unnoticeable this animation is.

note: the animate that happens to parallax on shuttles is still in place, it was handled separately.