bugfix: walking on tables (#4508)
* bugfix: walking on tables

* bugfix: walking on tables

* bugfix: walking on tables

* bugfix: walking on tables

* add: climbing animation

Added climbing animation when climbing onto a structures.

* bugfix: climbing in clown shoes

Fixed animation of clown boots when on structures

* add: fall animation

Added fall animation if the mob was above the surface.

* add: hopping animation

Added hopping animation when a mob climbs over a fence and small bugfix.

* add: hopping animation

Added hopping animation when a mob climbs over a railing and small bugfix.

* bugfix: disables climb animation

Disables the climbing animation on fences with a medium hole.
The time of climbing through the hole is increased by 3 times (values taken from define).

* bugfix: walking on tables

Fixed the possibility of walking on tables.
Added the feature of climbing on the skeleton bar.
Now the table can be accessed from any structures that can be climbed.

* Update structures.dm

* Update tables_racks.dm

* bugfix: walking on tables

* bugfix: walking on tables

* bugfix: walking on tables

* bugfix: disables climb animation

* add: hopping animation

* bugfix: walking on tables

* bugfix: walking on tables

Fix drop doll on pulling action and destroy action.

* bugfix: move stuctures with mob

* refactor: changed 'usr' to 'user'

* refactor: changed 'usr' to 'user'

* refactor: changed 'usr' to 'user'

* refactor: changed 'usr' to 'user'

* bugfix: moving a tray table with something on it

* bugfix: table drop damage for anthropomorphs only

Животные теперь не умирают сидя на столе.

* bugfix: multi-person fall

Co-authored-by: Rerik007 <120549107+Rerik007@users.noreply.github.com>

* refactor: change <span> to proc

* bugfix: Oops!

* refactor: usr to user

* refactor: standardization

Co-authored-by: BeebBeebBoob <beebbeebboob@mail.ru>

---------

Co-authored-by: Rerik007 <120549107+Rerik007@users.noreply.github.com>
Co-authored-by: BeebBeebBoob <beebbeebboob@mail.ru>