Last item in a long list should have a comma.
Last item in a long list should have a comma. Last item in a long list should have a comma. Last item in a long list should have a comma. Last item in a long list should have a comma. Last item in a long list should have a comma. Last item in a long list should have a comma. Last item in a long list should have a comma. Last item in a long list should have a comma. Last item in a long list should have a comma.